Datenbanken Updaten nur noch wenn eine Verbindung möglich ist
Verbindung wird alle 15 Sekunden überprüft.
Falls keine Verbindung möglich ist wird die vorherige Zeit ^2 genommen und gewartet.
Wenn eine Verbindung länger als 60Minuten nicht möglich war, wird der Thread beendet und somit wird bis zum nächsten Neustart keine Verbindung mehr Probiert

+++ BugHunter/BugHunter/Database.cs
@@ -13,8 +13,7 @@ namespace ProjectWhitespace
         /// <param name="game">Game Objekt um auf Einstellungen und Highscore zugreifen zu können</param>
         public static void UpdateDatabaseThread(Game1 game)
         {
-
-            String connString = "Server=" + Settings.host + ";Database=" + Settings.database
+            string connString = "Server=" + Settings.host + ";Database=" + Settings.database
                  + ";port=" + Settings.port + ";User Id=" + Settings.username + ";password=" + Settings.password;
 
             MySqlConnection connection = new MySqlConnection(connString);
@@ -23,82 +22,86 @@ namespace ProjectWhitespace
 
             while (true)
             {
-                try
+                if (game.settings.HasInternetConnection)
                 {
-                    if (connection.State != System.Data.ConnectionState.Open)
+
+                    try
                     {
-                        // Verbindung muss erst aufgebaut werden
-                        game.logger.Log("Datenbankverbindung wird aufgebaut", Thread.CurrentThread.Name, "Debug");
-                        connection.Open();
-
-                    }
-
-                    if (connection.State == System.Data.ConnectionState.Open)
-                    {
-                        // Datenbankverbindung steht
-                        game.logger.Log("Datenbankverbindung steht", Thread.CurrentThread.Name, "Debug");
-
-                        string query = "select `GlobalHighscore`.`userid`,`GlobalHighscore`.`score` from `GlobalHighscore`";
-                        command = new MySqlCommand(query);
-                        command.Connection = connection;
-
-                        // select rückgabe auslesen
-                        reader = command.ExecuteReader();
-
-                        bool GuidExists = false;
-
-                        while (reader.Read())
+                        if (connection.State != System.Data.ConnectionState.Open)
                         {
-                            // guid in datenbank gefunden
-                            if (reader.GetString(0).Equals(game.settings.GUID))
-                            {
-                                GuidExists = true;
-                                break;
-                            }
+                            // Verbindung muss erst aufgebaut werden
+                            game.logger.Log("Datenbankverbindung wird aufgebaut", Thread.CurrentThread.Name, "Debug");
+                            connection.Open();
+
                         }
 
-                        reader.Close();
-
-                        if (GuidExists)
+                        if (connection.State == System.Data.ConnectionState.Open)
                         {
-                            game.logger.Log("GUID war vorhanden. Eintrag wird upgedated", Thread.CurrentThread.Name, "Debug");
-                            // Datenbankeintrag wird upgedated
-                            command.CommandText =
-                                "UPDATE `GlobalHighscore` SET `Name` = '" + game.settings.UserName +
-                                "', `Score` = '" + game.gameStats.HighScore +
-                                "', `DateTime` = '" +
-                                DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss") +
-                                "' WHERE `GlobalHighscore`.`UserID` = '" +
-                                game.settings.GUID + "'";
-                            command.ExecuteNonQuery();
-                        }
-                        else
-                        {
-                            game.logger.Log("GUID nicht gefunden. Neuer Eintrag wird erstellt", Thread.CurrentThread.Name, "Debug");
-                            // Kein Eintrag gefunden, wodurch ein neuer erstellt wird
-                            command = new MySqlCommand("INSERT INTO `GlobalHighscore` (`UserID`, `Name`, `Score`, `DateTime`, `IPAddress`) VALUES('" +
-                                game.settings.GUID + "', '" +
-                                game.settings.UserName + "', '" +
-                                game.gameStats.HighScore + "', '" +
-                            DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss") +
-                            "', 'UNUSED');");
+                            // Datenbankverbindung steht
+                            game.logger.Log("Datenbankverbindung steht", Thread.CurrentThread.Name, "Debug");
 
+                            string query = "select `GlobalHighscore`.`userid`,`GlobalHighscore`.`score` from `GlobalHighscore`";
+                            command = new MySqlCommand(query);
                             command.Connection = connection;
 
-                            command.ExecuteNonQuery();
-                        }
+                            // select rückgabe auslesen
+                            reader = command.ExecuteReader();
 
+                            bool GuidExists = false;
+
+                            while (reader.Read())
+                            {
+                                // guid in datenbank gefunden
+                                if (reader.GetString(0).Equals(game.settings.GUID))
+                                {
+                                    GuidExists = true;
+                                    break;
+                                }
+                            }
+
+                            reader.Close();
+
+                            if (GuidExists)
+                            {
+                                game.logger.Log("GUID war vorhanden. Eintrag wird upgedated", Thread.CurrentThread.Name, "Debug");
+                                // Datenbankeintrag wird upgedated
+                                command.CommandText =
+                                    "UPDATE `GlobalHighscore` SET `Name` = '" + game.settings.UserName +
+                                    "', `Score` = '" + game.gameStats.HighScore +
+                                    "', `DateTime` = '" +
+                                    DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss") +
+                                    "' WHERE `GlobalHighscore`.`UserID` = '" +
+                                    game.settings.GUID + "'";
+                                command.ExecuteNonQuery();
+                            }
+                            else
+                            {
+                                game.logger.Log("GUID nicht gefunden. Neuer Eintrag wird erstellt", Thread.CurrentThread.Name, "Debug");
+                                // Kein Eintrag gefunden, wodurch ein neuer erstellt wird
+                                command = new MySqlCommand("INSERT INTO `GlobalHighscore` (`UserID`, `Name`, `Score`, `DateTime`, `IPAddress`) VALUES('" +
+                                    game.settings.GUID + "', '" +
+                                    game.settings.UserName + "', '" +
+                                    game.gameStats.HighScore + "', '" +
+                                DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss") +
+                                "', 'UNUSED');");
+
+                                command.Connection = connection;
+
+                                command.ExecuteNonQuery();
+                            }
+
+                        }
                     }
-                }
-                catch (MySqlException e)
-                {
-                    Console.WriteLine(e.Message);
-                    game.logger.Log(e.Message, "Error");
-                }
-                finally
-                {
-                    connection.Close();
-                    game.logger.Log("Datenbankverbindung geschlossen", Thread.CurrentThread.Name, "Debug");
+                    catch (MySqlException e)
+                    {
+                        Console.WriteLine(e.Message);
+                        game.logger.Log(e.Message,Thread.CurrentThread.Name, "Error");
+                    }
+                    finally
+                    {
+                        connection.Close();
+                        game.logger.Log("Datenbankverbindung geschlossen", Thread.CurrentThread.Name, "Debug");
+                    }
                 }
 
                 // Datenbank wird alle 15 Sekunden upgedated
@@ -109,7 +112,7 @@ namespace ProjectWhitespace
                 catch (ThreadInterruptedException e)
                 {
                     Console.WriteLine(e.Message);
-                    game.logger.Log("Database Update Thread beendet", Thread.CurrentThread.Name, "Debug");
+                    game.logger.Log("Thread beendet", Thread.CurrentThread.Name, "Debug");
                     break;
                 }
             }
@@ -125,55 +128,57 @@ namespace ProjectWhitespace
 
             while (true)
             {
-                try
+                if (game.settings.HasInternetConnection)
                 {
-                    if (connection.State != System.Data.ConnectionState.Open)
+                    try
                     {
-                        // Verbindung muss erst aufgebaut werden
-                        game.logger.Log("Datenbankverbindung wird aufgebaut", Thread.CurrentThread.Name, "Debug");
-                        connection.Open();
-
-                    }
-
-                    if (connection.State == System.Data.ConnectionState.Open)
-                    {
-                        // Datenbankverbindung steht
-                        game.logger.Log("Datenbankverbindung steht", Thread.CurrentThread.Name, "Debug");
-
-                        string query = "SELECT* FROM `GlobalHighscore` ORDER BY `Score` DESC";
-                        command = new MySqlCommand(query);
-                        command.Connection = connection;
-
-                        // select rückgabe auslesen
-                        reader = command.ExecuteReader();
-
-                        game.gameStats.Top10Names.RemoveRange(0, game.gameStats.Top10Names.Count);
-                        game.gameStats.Top10Score.RemoveRange(0, game.gameStats.Top10Score.Count);
-
-                        for (int i = 0; i <= 10 && reader.Read(); i++)
+                        if (connection.State != System.Data.ConnectionState.Open)
                         {
-                            // guid in datenbank gefunden
-                            if (!string.IsNullOrEmpty(reader.GetString(0)))
-                            {
-                                game.gameStats.Top10Names.Add(reader.GetString(1));
-                                game.gameStats.Top10Score.Add(reader.GetInt32(2));
-                            }
+                            // Verbindung muss erst aufgebaut werden
+                            game.logger.Log("Datenbankverbindung wird aufgebaut", Thread.CurrentThread.Name, "Debug");
+                            connection.Open();
+
                         }
 
-                        reader.Close();
+                        if (connection.State == System.Data.ConnectionState.Open)
+                        {
+                            // Datenbankverbindung steht
+                            game.logger.Log("Datenbankverbindung steht", Thread.CurrentThread.Name, "Debug");
+
+                            string query = "SELECT* FROM `GlobalHighscore` ORDER BY `Score` DESC";
+                            command = new MySqlCommand(query);
+                            command.Connection = connection;
+
+                            // select rückgabe auslesen
+                            reader = command.ExecuteReader();
+
+                            game.gameStats.Top10Names.RemoveRange(0, game.gameStats.Top10Names.Count);
+                            game.gameStats.Top10Score.RemoveRange(0, game.gameStats.Top10Score.Count);
+
+                            for (int i = 0; i <= 10 && reader.Read(); i++)
+                            {
+                                // guid in datenbank gefunden
+                                if (!string.IsNullOrEmpty(reader.GetString(0)))
+                                {
+                                    game.gameStats.Top10Names.Add(reader.GetString(1));
+                                    game.gameStats.Top10Score.Add(reader.GetInt32(2));
+                                }
+                            }
+
+                            reader.Close();
+                        }
+                    }
+                    catch (MySqlException e)
+                    {
+                        Console.WriteLine(e.Message);
+                        game.logger.Log(e.Message,Thread.CurrentThread.Name, "Error");
+                    }
+                    finally
+                    {
+                        connection?.Close();
+                        game.logger.Log("Datenbankverbindung geschlossen", Thread.CurrentThread.Name, "Debug");
                     }
                 }
-                catch (MySqlException e)
-                {
-                    Console.WriteLine(e.Message);
-                    game.logger.Log(e.Message, "Error");
-                }
-                finally
-                {
-                    connection?.Close();
-                    game.logger.Log("Datenbankverbindung geschlossen", Thread.CurrentThread.Name, "Debug");
-                }
-
                 try
                 {
                     Thread.Sleep(60000);
@@ -197,62 +202,65 @@ namespace ProjectWhitespace
 
             while (true)
             {
-                try
+                if (game.settings.HasInternetConnection)
                 {
-                    if (connection.State != System.Data.ConnectionState.Open)
+                    try
                     {
-                        // Verbindung muss erst aufgebaut werden
-                        game.logger.Log("Datenbankverbindung wird aufgebaut", Thread.CurrentThread.Name, "Debug");
-                        connection.Open();
+                        if (connection.State != System.Data.ConnectionState.Open)
+                        {
+                            // Verbindung muss erst aufgebaut werden
+                            game.logger.Log("Datenbankverbindung wird aufgebaut", Thread.CurrentThread.Name, "Debug");
+                            connection.Open();
 
+                        }
+
+                        if (connection.State == System.Data.ConnectionState.Open)
+                        {
+                            // Datenbankverbindung steht
+                            game.logger.Log("Datenbankverbindung steht", Thread.CurrentThread.Name, "Debug");
+
+                            string query = "SELECT * FROM `GlobalScore`";
+                            command = new MySqlCommand(query);
+                            command.Connection = connection;
+
+                            // select rückgabe auslesen
+                            reader = command.ExecuteReader();
+
+                            reader.Read();
+
+                            game.gameStats.GlobalKilledEnemies = reader.GetUInt32(1);
+                            game.gameStats.GlobalCollectedPowerups = reader.GetUInt32(2);
+                            game.gameStats.GlobalAnzahlSchuesse = reader.GetUInt64(3);
+                            game.gameStats.GlobalAnzahlTreffer = reader.GetUInt64(4);
+                            game.gameStats.GlobalAnzahlTode = reader.GetUInt32(5);
+
+                            reader.Close();
+
+                            // Globale Anzahl an Spielern aus Datenbank lesen
+                            query = "SELECT COUNT(*) FROM GlobalHighscore;";
+                            command = new MySqlCommand(query);
+                            command.Connection = connection;
+
+                            // select rückgabe auslesen
+                            reader = command.ExecuteReader();
+
+                            reader.Read();
+
+                            game.gameStats.GlobalPlayerAmount = reader.GetUInt32(0);
+
+                            reader.Close();
+                        }
                     }
-
-                    if (connection.State == System.Data.ConnectionState.Open)
+                    catch (MySqlException e)
                     {
-                        // Datenbankverbindung steht
-                        game.logger.Log("Datenbankverbindung steht", Thread.CurrentThread.Name, "Debug");
-
-                        string query = "SELECT * FROM `GlobalScore`";
-                        command = new MySqlCommand(query);
-                        command.Connection = connection;
-
-                        // select rückgabe auslesen
-                        reader = command.ExecuteReader();
-
-                        reader.Read();
-
-                        game.gameStats.GlobalKilledEnemies = reader.GetUInt32(1);
-                        game.gameStats.GlobalCollectedPowerups = reader.GetUInt32(2);
-                        game.gameStats.GlobalAnzahlSchuesse = reader.GetUInt64(3);
-                        game.gameStats.GlobalAnzahlTreffer = reader.GetUInt64(4);
-                        game.gameStats.GlobalAnzahlTode = reader.GetUInt32(5);
-
-                        reader.Close();
-
-                        // Globale Anzahl an Spielern aus Datenbank lesen
-                        query = "SELECT COUNT(*) FROM GlobalHighscore;";
-                        command = new MySqlCommand(query);
-                        command.Connection = connection;
-
-                        // select rückgabe auslesen
-                        reader = command.ExecuteReader();
-
-                        reader.Read();
-
-                        game.gameStats.GlobalPlayerAmount = reader.GetUInt32(0);
-
-                        reader.Close();
+                        Console.WriteLine(e.Message);
+                        game.logger.Log(e.Message,Thread.CurrentThread.Name, "Error");
                     }
-                }
-                catch (MySqlException e)
-                {
-                    Console.WriteLine(e.Message);
-                    game.logger.Log(e.Message, "Error");
-                }
-                finally
-                {
-                    connection?.Close();
-                    game.logger.Log("Datenbankverbindung geschlossen", Thread.CurrentThread.Name, "Debug");
+                    finally
+                    {
+                        connection?.Close();
+                        game.logger.Log("Datenbankverbindung geschlossen", Thread.CurrentThread.Name, "Debug");
+                    }
                 }
 
                 try
